backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
requests==2.32.3
python-dotenv==1.0.1
